backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pydantic==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0